+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - Added commands for executing C# expressions or statements at runtime.
 - Added exception handling for command callbacks.
+- Added a reminder in the devconsole command that the console is disabled by default in release builds.
 - Changed DevConsole.InvokeCoroutine method to return the Coroutine instance.
 - Fixed Unity logs from other threads not showing in the developer console log.